requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
gspread>=6.0.0
requests>=2.31.0
pyyaml>=6.0.1
msal>=1.28.0
